pyproject.toml
[project]
name = "openai-agents-sdk-class-02"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Zain Ali", email = "za24m04@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.4.400",
    "openai-agents>=0.0.7",
]

[project.scripts]
openai-agents-sdk-class-02 = "openai_agents_sdk_class_02:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
